Refactor: simpler variables for DB
