+++ DEPLOYMENT.md
@@ -12,29 +12,29 @@ To run new deployment, first create environment groups for secrets: `backend/mai
 For every backend there should be keys:
 
 ```
-NEAR_READ_ONLY_INDEXER_DATABASE_HOST
-NEAR_READ_ONLY_INDEXER_DATABASE_NAME
-NEAR_READ_ONLY_INDEXER_DATABASE_USERNAME
-NEAR_READ_ONLY_INDEXER_DATABASE_PASSWORD
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_INDEXER__HOST
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_INDEXER__DATABASE
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_INDEXER__USER
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_INDEXER__PASSWORD
 
-NEAR_READ_ONLY_ANALYTICS_DATABASE_HOST
-NEAR_READ_ONLY_ANALYTICS_DATABASE_NAME
-NEAR_READ_ONLY_ANALYTICS_DATABASE_USERNAME
-NEAR_READ_ONLY_ANALYTICS_DATABASE_PASSWORD
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_ANALYTICS__HOST
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_ANALYTICS__DATABASE
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_ANALYTICS__USER
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_ANALYTICS__PASSWORD
 
-NEAR_READ_ONLY_TELEMETRY_DATABASE_HOST
-NEAR_READ_ONLY_TELEMETRY_DATABASE_NAME
-NEAR_READ_ONLY_TELEMETRY_DATABASE_USERNAME
-NEAR_READ_ONLY_TELEMETRY_DATABASE_PASSWORD
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_TELEMETRY__HOST
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_TELEMETRY__DATABASE
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_TELEMETRY__USER
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_TELEMETRY__PASSWORD
 ```
 
 In case backend should expect node telemetry, there should be keys as well:
 
 ```
-NEAR_WRITE_ONLY_TELEMETRY_DATABASE_HOST
-NEAR_WRITE_ONLY_TELEMETRY_DATABASE_NAME
-NEAR_WRITE_ONLY_TELEMETRY_DATABASE_USERNAME
-NEAR_WRITE_ONLY_TELEMETRY_DATABASE_PASSWORD
+NEAR_EXPLORER_CONFIG__DB__WRITE_ONLY_TELEMETRY__HOST
+NEAR_EXPLORER_CONFIG__DB__WRITE_ONLY_TELEMETRY__DATABASE
+NEAR_EXPLORER_CONFIG__DB__WRITE_ONLY_TELEMETRY__USER
+NEAR_EXPLORER_CONFIG__DB__WRITE_ONLY_TELEMETRY__PASSWORD
 ```
 
 ### Blueprint
