+++ backend/src/config.ts
@@ -47,6 +47,35 @@ export const config = merge(
     },
 
     port: 10000,
+
+    db: {
+      readOnlyIndexer: {
+        host: "",
+        database: "",
+        user: "",
+        password: "",
+      },
+      readOnlyAnalytics: {
+        host: "",
+        database: "",
+        user: "",
+        password: "",
+      },
+      readOnlyTelemetry: {
+        host: "",
+        database: "",
+        user: "",
+        password: "",
+      },
+      writeOnlyTelemetry: {
+        host: "",
+        database: "",
+        user: "",
+        password: "",
+        min: 0,
+        max: 15,
+      },
+    },
   },
   getOverrides("NEAR_EXPLORER_CONFIG")
 );
