+++ backend/src/database/databases.ts
@@ -1,10 +1,10 @@
 import { Kysely, PostgresDialect, PostgresDialectConfig } from "kysely";
 import { Pool } from "pg";
-import { databaseConfigs } from "./configs";
+import { config } from "../config";
 
-import * as Indexer from "./models/readOnlyIndexerDatabase";
-import * as Telemetry from "./models/readOnlyTelemetryDatabase";
-import * as Analytics from "./models/readOnlyAnalyticsDatabase";
+import * as Indexer from "./models/readOnlyIndexer";
+import * as Telemetry from "./models/readOnlyTelemetry";
+import * as Analytics from "./models/readOnlyAnalytics";
 
 const getPgPool = (config: PostgresDialectConfig): Pool => {
   const pool = new Pool(config);
@@ -24,25 +24,22 @@ const getKysely = <T>(config: PostgresDialectConfig): Kysely<T> =>
     dialect: new PostgresDialect(getPgPool(config)),
   });
 
-export const telemetryWriteDatabase = databaseConfigs.writeOnlyTelemetryDatabase
-  .host
-  ? getKysely<Telemetry.ModelTypeMap>(
-      databaseConfigs.writeOnlyTelemetryDatabase
-    )
+export const telemetryWriteDatabase = config.db.writeOnlyTelemetry.host
+  ? getKysely<Telemetry.ModelTypeMap>(config.db.writeOnlyTelemetry)
   : null;
 
 export const telemetryDatabase = getKysely<Telemetry.ModelTypeMap>(
-  databaseConfigs.readOnlyTelemetryDatabase
+  config.db.readOnlyTelemetry
 );
 
 export const indexerDatabase = getKysely<Indexer.ModelTypeMap>(
-  databaseConfigs.readOnlyIndexerDatabase
+  config.db.readOnlyIndexer
 );
 
 export const analyticsDatabase = getKysely<Analytics.ModelTypeMap>(
-  databaseConfigs.readOnlyAnalyticsDatabase
+  config.db.readOnlyAnalytics
 );
 
-export const extraPool = getPgPool(databaseConfigs.writeOnlyTelemetryDatabase);
+export const extraPool = getPgPool(config.db.writeOnlyTelemetry);
 
 export { Indexer, Analytics, Telemetry };
