+++ backend/src/database/models/readOnlyAnalytics/daily-accounts-added-per-ecosystem-entity.ts
