+++ backend/src/database/models/readOnlyAnalytics/daily-deleted-accounts-count.ts
