+++ backend/src/database/models/readOnlyAnalytics/daily-deposit-amount.ts
