+++ backend/src/database/models/readOnlyAnalytics/daily-gas-used.ts
