+++ backend/src/database/models/readOnlyAnalytics/daily-new-accounts-count.ts
