+++ backend/src/database/models/readOnlyAnalytics/daily-new-accounts-per-ecosystem-entity-count.ts
