+++ backend/src/database/models/readOnlyAnalytics/daily-new-contracts-count.ts
