+++ backend/src/database/models/readOnlyAnalytics/daily-receipts-per-contract-count.ts
