+++ backend/src/database/models/readOnlyAnalytics/daily-tokens-spent-on-fees.ts
