+++ backend/src/database/models/readOnlyAnalytics/transactions-with-converted-receipts.ts
