+++ backend/src/database/models/readOnlyIndexer/action-receipt-input-data.ts
