+++ backend/src/database/models/readOnlyIndexer/action-receipt-output-data.ts
