+++ backend/src/database/models/readOnlyIndexer/aggregated-circulating-supply.ts
