+++ backend/src/database/models/readOnlyIndexer/assets-fungible-token-events.ts
