+++ backend/src/database/models/readOnlyIndexer/assets-non-fungible-token-events.ts
