+++ backend/src/database/models/readOnlyIndexer/receipt-kind.ts
