+++ backend/src/database/models/readOnlyTelemetry/nodes.ts
