+++ testnet.env
@@ -6,22 +6,22 @@ NEAR_EXPLORER_CONFIG__ARCHIVAL_RPC_URL=https://archival-rpc.testnet.near.org
 # Learn more about running your own Indexer for Explorer here: https://github.com/near/near-indexer-for-explorer
 
 # NOTE: These are public services that operate under fair use conditions. Do not abuse them.
-NEAR_READ_ONLY_INDEXER_DATABASE_HOST=testnet.db.explorer.indexer.near.dev
-NEAR_READ_ONLY_INDEXER_DATABASE_NAME=testnet_explorer
-NEAR_READ_ONLY_INDEXER_DATABASE_USERNAME=public_readonly
-NEAR_READ_ONLY_INDEXER_DATABASE_PASSWORD=nearprotocol
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_INDEXER__HOST=testnet.db.explorer.indexer.near.dev
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_INDEXER__DATABASE=testnet_explorer
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_INDEXER__USER=public_readonly
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_INDEXER__PASSWORD=nearprotocol
 
-NEAR_READ_ONLY_ANALYTICS_DATABASE_HOST=35.241.197.241
-NEAR_READ_ONLY_ANALYTICS_DATABASE_NAME=indexer_analytics_testnet
-NEAR_READ_ONLY_ANALYTICS_DATABASE_USERNAME=public_readonly
-NEAR_READ_ONLY_ANALYTICS_DATABASE_PASSWORD=nearprotocol
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_ANALYTICS__HOST=35.241.197.241
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_ANALYTICS__DATABASE=indexer_analytics_testnet
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_ANALYTICS__USER=public_readonly
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_ANALYTICS__PASSWORD=nearprotocol
 
-NEAR_READ_ONLY_TELEMETRY_DATABASE_HOST=35.241.197.241
-NEAR_READ_ONLY_TELEMETRY_DATABASE_NAME=telemetry_testnet
-NEAR_READ_ONLY_TELEMETRY_DATABASE_USERNAME=public_readonly
-NEAR_READ_ONLY_TELEMETRY_DATABASE_PASSWORD=nearprotocol
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_TELEMETRY__HOST=35.241.197.241
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_TELEMETRY__DATABASE=telemetry_testnet
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_TELEMETRY__USER=public_readonly
+NEAR_EXPLORER_CONFIG__DB__READ_ONLY_TELEMETRY__PASSWORD=nearprotocol
 
-NEAR_WRITE_ONLY_TELEMETRY_DATABASE_HOST=
-NEAR_WRITE_ONLY_TELEMETRY_DATABASE_NAME=telemetry_testnet
-NEAR_WRITE_ONLY_TELEMETRY_DATABASE_USERNAME=
-NEAR_WRITE_ONLY_TELEMETRY_DATABASE_PASSWORD=
+NEAR_EXPLORER_CONFIG__DB__WRITE_ONLY_TELEMETRY__HOST=
+NEAR_EXPLORER_CONFIG__DB__WRITE_ONLY_TELEMETRY__DATABASE=telemetry_testnet
+NEAR_EXPLORER_CONFIG__DB__WRITE_ONLY_TELEMETRY__USER=
+NEAR_EXPLORER_CONFIG__DB__WRITE_ONLY_TELEMETRY__PASSWORD=
